Change path of CacheUpdater.
