+++ paddlehub/common/downloader.py
@@ -27,9 +27,6 @@ import requests
 import tarfile
 
 from paddlehub.common import utils
-from paddlehub.common.logger import logger
-from paddlehub.common import tmp_dir
-from paddlehub.common.hub_server import CacheUpdater
 import paddlehub as hub
 
 __all__ = ['Downloader', 'progress']
@@ -177,36 +174,3 @@ class ServerConnectionError(Exception):
         tips = 'Can\'t connect to Hub Server:{}'.format(
             hub.HubServer().server_url[0])
         return tips
-
-
-def download(name,
-             save_path,
-             version=None,
-             decompress=True,
-             resource_type='Model',
-             extra=None):
-    file = os.path.join(save_path, name)
-    file = os.path.realpath(file)
-    if os.path.exists(file):
-        return
-
-    if not hub.HubServer()._server_check():
-        raise ServerConnectionError
-
-    search_result = hub.HubServer().get_resource_url(
-        name, resource_type=resource_type, version=version, extra=extra)
-
-    if not search_result:
-        raise ResourceNotFoundError(name, version)
-    CacheUpdater("hub_download", name, version).start()
-    url = search_result['url']
-
-    with tmp_dir() as _dir:
-        if not os.path.exists(save_path):
-            os.makedirs(save_path)
-        _, _, savefile = default_downloader.download_file(
-            url=url, save_path=_dir, print_progress=True)
-        if tarfile.is_tarfile(savefile) and decompress:
-            _, _, savefile = default_downloader.uncompress(
-                file=savefile, print_progress=True)
-        shutil.move(savefile, file)
